requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
bcrypt==3.1.6
billiard==3.5.0.5
cached-property==1.5.1
celery==4.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cli-helpers==1.2.0
Click==7.0
configobj==5.0.6
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.3
cryptography==2.6.1
Django==2.1.7
django-bootstrap4==0.0.7
django-concurrency==2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-guardian==1.5.0
django-online-users==0.3
django-postgres-extensions==0.9.3
django-registration==3.0
django-silk==3.0.1
djangorestframework==3.9.0
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.1
flake8==3.6.0
gprof2dot==2016.10.13
humanize==0.5.1
idna==2.7
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
jsonschema==2.6.0
kombu==4.2.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==7.0.0
paramiko==2.4.2
pgcli==2.1.0
pgspecial==1.11.5
pluggy==0.9.0
prompt-toolkit==2.0.9
psycopg2==2.7.6.1
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
PyNaCl==1.3.0
pytest==4.4.0
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-docker-compose==1.0.1
python-dateutil==2.8.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
setproctitle==1.1.10
six==1.12.0
sqlparse==0.2.4
tabulate==0.8.3
terminaltables==3.1.0
text-unidecode==1.2
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.56.0
